backend/requirements.txt
fastapi
uvicorn[standard]
httpx
astroquery
astropy
numpy
pandas
scikit-learn
xgboost
diskcache
python-dotenv
joblib
pydantic